requirements.txt
# Django + DRF stack
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.14
qrcode>=7.4
